Updated the blog delete function to only delete blog if the op is logged in

+++ osa4/blogilista/controllers/blogs.js
@@ -11,7 +11,6 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 
-
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
@@ -34,6 +33,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    const user = await User.findById(decodedToken.id)
+
+    if ( blog.user.toString() != user.id )
+        return response.status(401).json({ error: 'user has no rights to this blog' })
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
